Handle invalid and valid packets

+++ src/Ground-Terminal-System/ChecksumValidator.cs
@@ -1,12 +1,18 @@
+/*
+ * File Name    : ChecksumValidator.cs
+ * Description  : This is the class for validating checksums of incoming telemetry data packets.
+ * Author       : Andrei Haboc
+ * Last Modified: November 28, 2025
+ */
 namespace GroundTerminalSystem
 {
     public class ChecksumValidator
     {
 
-        public bool Validate(float altitude, float pitch, float bank, int receivedChecksum)
+        public bool Validate(double altitude, double pitch, double bank, int receivedChecksum)
         {
             // Checksum = (ALT + Pitch + Bank) / 3
-            float calculation = (altitude + pitch + bank) / 3;
+            double calculation = ((altitude + pitch + bank) / 3);
 
             int calculatedChecksum = (int)calculation;
